chore: refactor with OCRInput
