+++ JWords/macView/OCR/OCRView.swift
@@ -10,89 +10,27 @@ import ComposableArchitecture
 
 struct OCR: ReducerProtocol {
     struct State: Equatable {
-        
-        enum Mode: Equatable {
-            case insert
-            case addExist(existing: StudyUnit)
-        }
-        
+
         // OCR
         var image: InputImageType?
         var getImageButtons = GetImageForOCR.State()
         var koreanOcrResult: [OCRResult] = []
         var japaneseOcrResult: [OCRResult] = []
         
-        // input
-        var mode: Mode = .insert
-        var selectSet = SelectStudySet.State(pickerName: "단어장")
-        var kanjiString: String = ""
-        var meaningString: String = ""
-        var huriText: EditHuriganaText.State?
+        var input = OCRInput.State()
         var alert: AlertState<Action>?
         
-        var huriString: String? {
-            huriText?.hurigana
-        }
-        
-        fileprivate mutating func clearUserInput() {
-            mode = .insert
-            kanjiString = ""
-            meaningString = ""
-            huriText = nil
-            #if os(iOS)
-            dismissKeyBoard()
-            #endif
-        }
-        
-        fileprivate mutating func setExistAlert() {
-            let alreadyExist = kanjiString
+        fileprivate mutating func setAlert(_ type: OCRInput.AlertType) {
             alert = AlertState<Action> {
-                TextState("표제어 중복")
+                TextState(type.title)
             } actions: {
                 ButtonState(role: .none) {
                     TextState("확인")
                 }
             } message: {
-                TextState("\(alreadyExist)와 동일한 단어가 존재합니다")
+                TextState(type.message)
             }
         }
-        
-        fileprivate mutating func setNoSetAlert() {
-            alert = AlertState<Action> {
-                TextState("단어장 선택 안됨")
-            } actions: {
-                ButtonState(role: .none) {
-                    TextState("확인")
-                }
-            } message: {
-                TextState("단어장을 선택해주세요")
-            }
-        }
-        
-        fileprivate mutating func setNoHuriAlert() {
-            alert = AlertState<Action> {
-                TextState("가나 변환 필요")
-            } actions: {
-                ButtonState(role: .none) {
-                    TextState("확인")
-                }
-            } message: {
-                TextState("저장하기 전에 후리가나로 변환해야 합니다.")
-            }
-        }
-        
-        fileprivate mutating func setNoMeaningAlert() {
-            alert = AlertState<Action> {
-                TextState("뜻 없음")
-            } actions: {
-                ButtonState(role: .none) {
-                    TextState("확인")
-                }
-            } message: {
-                TextState("뜻이 없습니다.")
-            }
-        }
-        
     }
     
     enum Action: Equatable {
@@ -102,19 +40,11 @@ struct OCR: ReducerProtocol {
         case koreanOcrResponse(TaskResult<[OCRResult]>)
         case japaneseOcrResponse(TaskResult<[OCRResult]>)
         case ocrTapped(lang: OCRLang, string: String)
-        
-        case selectSet(SelectStudySet.Action)
-        case updateKanjiString(String)
-        case updateMeaningString(String)
-        case convertButtonTapped
-        case revertButtonTapped
-        case huriText(EditHuriganaText.Action)
-        case alertDismissed
-        case saveButtonTapped
+        case input(OCRInput.Action)
+        case dismissAlert
     }
     
     @Dependency(\.pasteBoardClient) var pasteBoardClient
-    private let cd = CoreDataClient.shared
     
     
     var body: some ReducerProtocol<State, Action> {
@@ -154,92 +84,38 @@ struct OCR: ReducerProtocol {
             case .japaneseOcrResponse(.success(let results)):
                 state.japaneseOcrResult = results
                 return .none
-            case .updateKanjiString(let string):
-                state.kanjiString = string
-                return .none
-            case .updateMeaningString(let string):
-                state.meaningString = string
-                return .none
             case .ocrTapped(let lang, let string):
                 switch lang {
                 case .korean:
-                    state.meaningString = string
+                    state.input.meaningString = string
                 case .japanese:
-                    state.huriText = nil
-                    state.kanjiString = string
+                    state.input.huriText = nil
+                    state.input.kanjiString = string
                 }
                 return .none
-            case .convertButtonTapped:
-                let huri = HuriganaConverter.shared.convert(state.kanjiString)
-                state.huriText = EditHuriganaText.State(hurigana: huri)
-                let unit = try! cd.checkIfExist(huri)
-                if let unit = unit {
-                    state.mode = .addExist(existing: unit)
-                    state.meaningString = unit.meaningText ?? ""
-                    state.setExistAlert()
-                } else {
-                    state.mode = .insert
+            case .input(let action):
+                switch action {
+                case .showAlert(let type):
+                    state.setAlert(type)
+                    return .none
+                default:
+                    return .none
                 }
-                return .none
-            case .revertButtonTapped:
-                state.huriText = nil
-                state.mode = .insert
-                return .none
-            case .alertDismissed:
+            case .dismissAlert:
                 state.alert = nil
-                return .none
-            case .saveButtonTapped:
-                guard let set = state.selectSet.selectedSet else {
-                    state.setNoSetAlert()
-                    return .none
-                }
-                guard let huri = state.huriString else {
-                    state.setNoHuriAlert()
-                    return .none
-                }
-                guard !state.meaningString.isEmpty else {
-                    state.setNoMeaningAlert()
-                    return .none
-                }
-                switch state.mode {
-                case .insert:
-                    _ = try! cd.insertUnit(in: set,
-                                           type: .word,
-                                           kanjiText: huri,
-                                           meaningText: state.meaningString)
-                case .addExist(let unit):
-                    _ = try! cd.addExistingUnit(unit: unit,
-                                                meaningText: state.meaningString,
-                                                in: set)
-                }
-                state.clearUserInput()
-                state.selectSet.onUnitAdded()
                 return .none
             default:
                 return .none
             }
         }
-        .ifLet(\.huriText, action: /Action.huriText) {
-            EditHuriganaText()
-        }
-        Scope(state: \.selectSet, action: /Action.selectSet) {
-            SelectStudySet()
-        }
         Scope(state: \.getImageButtons, action: /Action.getImageButtons) {
             GetImageForOCR()
         }
+        Scope(state: \.input, action: /Action.input) {
+            OCRInput()
+        }
     }
-
 }
-
-// TODO: move this to somewhere proper
-
-fileprivate func dismissKeyBoard() {
-    #if os(iOS)
-    UIApplication.shared.sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
-    #endif
-}
-
 
 fileprivate func resizeImage(_ image: InputImageType, to newSize: CGSize) -> InputImageType? {
     #if os(iOS)
@@ -278,10 +154,6 @@ struct OCRView: View {
         WithViewStore(store, observe: { $0 }) { vs in
             ScrollView {
                 VStack {
-                    StudySetPicker(store: store.scope(
-                        state: \.selectSet,
-                        action: OCR.Action.selectSet)
-                    )
                     if let image = vs.image {
                         VStack {
                             OCRResultView(image: image, koreanResults: vs.koreanOcrResult, japaneseResults: vs.japaneseOcrResult) { vs.send(.ocrTapped(lang: $0, string: $1)) }
@@ -297,45 +169,10 @@ struct OCRView: View {
                         )
                         .padding(.vertical, 20)
                     }
-                    HStack {
-                        VStack {
-                            Text("単語")
-                            if vs.huriText != nil {
-                                IfLetStore(self.store.scope(state: \.huriText,
-                                                            action: OCR.Action.huriText)
-                                ) {
-                                    EditableHuriganaText(store: $0)
-                                }
-                                Button("수정") {
-                                    vs.send(.revertButtonTapped)
-                                }
-                            } else {
-                                TextEditor(text: vs.binding(get: \.kanjiString, send: OCR.Action.updateKanjiString))
-                                    .font(.system(size: 30))
-                                    .border(.black)
-                                    .frame(height: 100)
-                                Button("변환") {
-                                    vs.send(.convertButtonTapped)
-                                }
-                                .disabled(vs.kanjiString.isEmpty)
-                            }
-                            Spacer()
-                        }
-                        VStack {
-                            Text("意味")
-                            TextEditor(text: vs.binding(get: \.meaningString, send: OCR.Action.updateMeaningString))
-                                .font(.system(size: 30))
-                                .border(.black)
-                                .frame(height: 100)
-                            Spacer()
-                        }
-                    }
-                    Button(vs.mode == .insert
-                           ? "새 단어 추가"
-                           : "기존 단어 단어장에 추가"
-                    ) {
-                        vs.send(.saveButtonTapped)
-                    }
+                    OCRInputView(store: store.scope(
+                        state: \.input,
+                        action: OCR.Action.input)
+                    )
                 }
             }
             .padding(.bottom, 50)
@@ -355,7 +192,7 @@ struct OCRView: View {
             #endif
             .alert(
               self.store.scope(state: \.alert),
-              dismiss: .alertDismissed
+              dismiss: .dismissAlert
             )
         }
     }
